+++ web/api/views.py
@@ -55,10 +55,7 @@ repconf = Config("reporting")
 
 if repconf.mongodb.enabled:
     import pymongo
-
-    results_db = pymongo.MongoClient(
-        settings.MONGO_HOST, port=settings.MONGO_PORT, username=settings.MONGO_USER, password=settings.MONGO_PASS, authSource=settings.MONGO_DB
-    )[settings.MONGO_DB]
+    results_db = pymongo.MongoClient(settings.MONGO_HOST, port=settings.MONGO_PORT, username=settings.MONGO_USER, password=settings.MONGO_PASS, authSource=settings.MONGO_DB)[settings.MONGO_DB]
 
 es_as_db = False
 if repconf.elasticsearchdb.enabled and not repconf.elasticsearchdb.searchonly:
@@ -205,7 +202,6 @@ def tasks_create_file(request):
         # Parse potential POST options (see submission/views.py)
         quarantine = request.POST.get("quarantine", "")
         pcap = request.POST.get("pcap", "")
-
         unique = bool(request.POST.get("unique", False))
         static = request.POST.get("static", "")
         priority = force_int(request.POST.get("priority"))
@@ -217,10 +213,22 @@ def tasks_create_file(request):
                 options += ","
             options += "procmemdump=1,procdump=1"
 
+        details = {
+            "errors": [],
+            "request": request,
+            "task_id": [],
+            "url": False,
+            "params": {},
+            "headers": {},
+            "service": "tasks_create_file_API",
+            "fhash": False,
+            "options": options,
+            "only_extraction": False,
+        }
+
         task_ids_tmp = []
         task_machines = []
         vm_list = []
-        details = {}
         for vm in db.list_machines():
             vm_list.append(vm.label)
 
@@ -261,6 +269,7 @@ def tasks_create_file(request):
                 resp = {"error": True, "error_value": "File size exceeds API limit"}
                 return jsonize(resp, response=True)
             tmp_path = store_temp_file(sample.read(), sanitize_filename(sample.name))
+            details["path"] = tmp_path
             if unique and db.check_file_uniq(File(tmp_path).get_sha256()):
                 details["errors"].append({sample.name: "Not unique, as unique option set"})
                 continue
@@ -296,23 +305,7 @@ def tasks_create_file(request):
                     details["errors"].append({os.path.basename(tmp_path):"Error submitting file - bad file type"})
                     continue
             else:
-                content = get_file_content(tmp_path)
-                details = {
-                    "errors": [],
-                    "content": content,
-                    "request": request,
-                    "task_id": [],
-                    "url": False,
-                    "params": {},
-                    "headers": {},
-                    "service": "tasks_create_file_API",
-                    "path": tmp_path,
-                    "fhash": False,
-                    "options": options,
-                    "only_extraction": False,
-                    "task_machines": task_machines,
-                }
-
+                details["content"] = get_file_content(tmp_path)
                 status, task_ids_tmp = download_file(**details)
                 if status == "error":
                     details["errors"].append({os.path.basename(tmp_path): task_ids_tmp})
@@ -499,7 +492,6 @@ def tasks_create_dlnexec(request):
             "fhash": False,
             "options": options,
             "only_extraction": False,
-            "task_machines": task_machines,
         }
 
         status, task_ids_tmp = download_file(**details)
@@ -562,7 +554,7 @@ def tasks_vtdl(request):
         if opts:
             opt_apikey = opts.get("apikey", False)
 
-        if (not settings.VTDL_PRIV_KEY and not settings.VTDL_INTEL_KEY) or not settings.VTDL_PATH or not opt_apikey:
+        if settings.VTDL_KEY or not settings.VTDL_PATH or not opt_apikey:
             resp = {"error": True, "error_value": "You specified VirusTotal but must edit the file and specify your VTDL_KEY variable and VTDL_PATH base directory"}
             return jsonize(resp, response=True)
 
@@ -599,7 +591,6 @@ def tasks_vtdl(request):
             "fhash": False,
             "options": options,
             "only_extraction": False,
-            "task_machines": task_machines,
         }
 
         details = download_from_vt(hashes, details, opt_filename, settings)
@@ -1077,8 +1068,8 @@ def tasks_report(request, task_id, report_format="json"):
 
     elif report_format.lower() in bz_formats:
         bzf = bz_formats[report_format.lower()]
-        srcdir = os.path.join(CUCKOO_ROOT, "storage", "analyses", "%d" % task_id)
-        s = BytesWarning()
+        srcdir = os.path.join(CUCKOO_ROOT, "storage", "analyses", str(task_id))
+        s = BytesIO()
 
         # By default go for bz2 encoded tar files (for legacy reasons.)
         # tarmode = tar_formats.get("tar", "w:bz2")
@@ -1169,6 +1160,9 @@ def tasks_iocs(request, task_id, detail=None):
         if data["target"]["category"] == "file":
             del data["target"]["file"]["path"]
             del data["target"]["file"]["guest_paths"]
+            for x in data["target"]["file"]["yara"]:
+                for i in range(0, len(x["strings"])):
+                    x["strings"][i] = x["strings"][i].hex()
 
     data["network"] = {}
     if "network" in list(buf.keys()) and buf["network"]:
@@ -1828,11 +1822,13 @@ def tasks_payloadfiles(request, task_id):
     try:
         zippwd = settings.ZIP_PWD
     except AttributeError:
-        zippwd = "infected"
+        zippwd = b"infected"
 
     capepath = os.path.join(CUCKOO_ROOT, "storage", "analyses", task_id, "CAPE")
 
     if os.path.exists(capepath):
+        if not HAVE_PYZIPPER:
+            return jsonize({"error": True, "error_value": "Install pyzipper to be able to download files"}, response=True)
         mem_zip = BytesIO()
         with pyzipper.AESZipFile(mem_zip, 'w', compression=pyzipper.ZIP_LZMA, encryption=pyzipper.WZ_AES) as zf:
             zf.setpassword(zippwd)
@@ -1842,15 +1838,15 @@ def tasks_payloadfiles(request, task_id):
                     with open(filepath, "rb") as f:
                         zf.writestr(os.path.basename(filepath), f.read())
 
-        # ToDo
-        #resp = StreamingHttpResponse(FileWrapper(open(zip_file), 8192), content_type="application/zip")
-        resp = HttpResponse(mem_zip.getvalue(), ontent_type="application/zip")
-        resp["Content-Length"] = os.path.getsize(len(mem_zip))
-        resp["Content-Disposition"] = "attachment; filename=" + "cape_payloads_{}.zip".format(task_id)
+
+        mem_zip.seek(0)
+        resp = StreamingHttpResponse(mem_zip, content_type="application/zip")
+        #resp = HttpResponse(mem_zip.getvalue(), content_type="application/zip")
+        resp["Content-Length"] = len(mem_zip.getvalue())
+        resp["Content-Disposition"] = f"attachment; filename=cape_payloads_{task_id}.zip"
         return resp
     else:
-        resp = {"error": True, "error_value": "No CAPE file(s) for task {}.".format(task_id)}
-        return jsonize(resp, response=True)
+        return jsonize({"error": True, "error_value": f"No CAPE file(s) for task {task_id}."}, response=True)
 
 
 @ratelimit(key="ip", rate=my_rate_seconds, block=rateblock)
@@ -1877,7 +1873,10 @@ def tasks_procdumpfiles(request, task_id):
 
     procdumppath = os.path.join(CUCKOO_ROOT, "storage", "analyses", task_id, "procdump")
 
+    #ToDo check bad rturn
     if os.path.exists(procdumppath):
+        if not HAVE_PYZIPPER:
+            return jsonize({"error": True, "error_value": "Install pyzipper to be able to download files"}, response=True)
         mem_zip = BytesIO()
         with pyzipper.AESZipFile(mem_zip, 'w', compression=pyzipper.ZIP_LZMA, encryption=pyzipper.WZ_AES) as zf:
             zf.setpassword(zippwd)
@@ -1887,14 +1886,14 @@ def tasks_procdumpfiles(request, task_id):
                     with open(filepath, "rb") as f:
                         zf.writestr(os.path.basename(filepath), f.read())
 
-        #ToDo
-        #resp = StreamingHttpResponse(FileWrapper(open(zip_file), 8192), content_type="application/zip")
-        resp = HttpResponse(mem_zip.getvalue(), ontent_type="application/zip")
-        resp["Content-Length"] = os.path.getsize(len(mem_zip))
-        resp["Content-Disposition"] = "attachment; filename=" + "cape_payloads_{}.zip".format(task_id)
+        mem_zip.seek(0)
+        resp = StreamingHttpResponse(mem_zip, content_type="application/zip")
+        #resp = HttpResponse(mem_zip.getvalue(), content_type="application/zip")
+        resp["Content-Length"] = len(mem_zip.getvalue())
+        resp["Content-Disposition"] = f"attachment; filename=cape_payloads_{task_id}.zip"
         return resp
     else:
-        resp = {"error": True, "error_value": "No procdump file(s) for task {}.".format(task_id)}
+        resp = {"error": True, "error_value": f"No procdump file(s) for task {task_id}."}
         return jsonize(resp, response=True)
 
 
@@ -1942,11 +1941,17 @@ def tasks_config(request, task_id, cape_name=False):
                 # In case compress results processing module is not enabled
                 pass
             data = []
-            for cape in buf["CAPE"]:
-                if isinstance(cape, dict) and cape.get("cape_config"):
-                    if cape_name and cape.get("cape_name", "") == cape_name:
-                        return jsonize(cape["cape_config"], response=True)
-                    data.append(cape)
+            if not isinstance(buf["CAPE"], list) and buf["CAPE"].get("configs"):
+                if cape_name and buf["CAPE"]["configs"].get("cape_name", "") == cape_name:
+                    return jsonize({cape_name.lower(): buf["CAPE"]["configs"][cape_name]}, response=True)
+                data = buf["CAPE"]["configs"]
+            # ToDo remove in v3
+            elif buf["CAPE"]:
+                for cape in buf["CAPE"]:
+                    if isinstance(cape, dict) and cape.get("cape_config"):
+                        if cape_name and cape.get("cape_name", "") == cape_name:
+                            return jsonize(cape["cape_config"], response=True)
+                        data.append(cape)
             if data:
                 resp = {"error": False, "configs": data}
             else: