+++ web/web/settings.py
@@ -35,8 +35,7 @@ if not cfg.mongodb.get("enabled") and not cfg.elasticsearchdb.get("enabled"):
 if cfg.mongodb.get("enabled") and cfg.elasticsearchdb.get("enabled") and not cfg.elasticsearchdb.get("searchonly"):
     raise Exception("Both database backend reporting modules are enabled. Please only enable ElasticSearch or MongoDB.")
 
-# Enable Django authentication for website
-WEB_AUTHENTICATION = False
+WEB_AUTHENTICATION = web_cfg.web_auth.get("enabled", False)
 
 # Get connection options from reporting.conf.
 MONGO_HOST = cfg.mongodb.get("host", "127.0.0.1")
@@ -55,7 +54,9 @@ zip_cfg = aux_cfg.zipped_download
 
 URL_ANALYSIS = web_cfg.url_analysis.get("enabled", False)
 DLNEXEC = web_cfg.dlnexec.get("enabled", False)
-ZIP_PWD = zip_cfg.get("zip_pwd", "infected")
+ZIP_PWD = zip_cfg.get("zip_pwd", b"infected")
+if not isinstance(ZIP_PWD, bytes):
+    ZIP_PWD = ZIP_PWD.encode("utf-8")
 MOLOCH_BASE = moloch_cfg.get("base", None)
 MOLOCH_NODE = moloch_cfg.get("node", None)
 MOLOCH_ENABLED = moloch_cfg.get("enabled", False)